Rename cache to libcache
